New and improved setup.
This version does all the compiling and moving and making files first.
Then it generates a batch script with all the right commands and puts it
in the output same directory as the binary. Finally, it submits the
batch script from inside that directory so the output stays there too.